+++ app.R
@@ -12,7 +12,7 @@ source("./global.R")
 # UI ----
 ui <- fluidPage(
   # App title
-  titlePanel(title = "TCU trend analysis | V 0.0.2"),
+  titlePanel(title = "TCU trend analysis | V 0.0.3"),
   
   fluidRow(
     column(3,
@@ -42,22 +42,39 @@ server <- shinyServer(function(input, output, session) {
       mes <- validateFileDF(f.df)
       
       # Render files df
-      output$input_files_table <- DT::renderDataTable(f.df %>% select(-datapath, -size, -type), 
+      output$input_files_table <- DT::renderDataTable(f.df %>% select(name, date, eqp_no, iter), 
                                                       options = list(dom = 't'))
       
       if(!is.null(mes)) {
         print(mes)
         output$mes <- renderText(paste0("Incorrect file input: ",
                                         mes))
-        
+        return(F)
+      } 
+      
+      # Validate data
+      f.df <- validateData(f.df)
+      if(any(!f.df[, "nrow"])) mes <- "file does not contain output from 9 probes"
+      if(any(!f.df[, "column_names"])) mes <- "file does not contain all column names"
+      if(any(!f.df[, "column_types"])) mes <- "column types not as expected"
+      if(any(!f.df[, "complete"])) mes <- "file contains empty cells"
+      
+      output$input_files_table <- DT::renderDataTable(f.df %>% select(name, date, eqp_no, iter, nrow, column_names, column_types, complete), 
+                                                      options = list(dom = 't'))
+      
+      if(!is.null(mes)) {
+        print(mes)
+        output$mes <- renderText(paste0("Incorrect data input: ",
+                                        mes))
         return(F)
       } else{
+        # Define EQP_NO for plot title
         EQP_NO <<- f.df$eqp_no[1]
         
         # Preprocess data
         WDF <- preprocessData(f.df)
         
-        # Gather data frame and change date column to posix
+        # Gather data frame and change date column to posix (for plotting the dates)
         LDF <<- WDF %>% gather(variable, value, -date, -Channel) %>% 
           mutate(date = as.POSIXct(date))
         return(T)
@@ -68,9 +85,9 @@ server <- shinyServer(function(input, output, session) {
   outputOptions(output, 'file_uploaded', suspendWhenHidden=FALSE)
   
   # Sidebar ----
-  # Extract mm selected by mm radar
-  mm <- reactive({
-    return(input$radio_mm)
+  # Extract limit selected by limit radar
+  limit <- reactive({
+    return(input$radio_limit)
   })
   
   observeEvent({
@@ -80,15 +97,16 @@ server <- shinyServer(function(input, output, session) {
     output$download_plot = downloadHandler(
       filename = function() {paste0(Sys.Date(), "_", EQP_NO, ".png")},
       content = function(file) {
-        ggsave(file, width = 10, height = 6, units = "in", plot = plotSubTrend(LDF, input$radio_mm[1], size = 10), device = "png")
+        ggsave(file, width = 10, height = 6, units = "in", plot = plotSubTrend(LDF, input$radio_limit[1], size = 10), device = "png")
       }
     )
   })
   
   output$probe_overview <- renderImage({
+    # Define outfile
     outfile <- tempfile(fileext = '.png')
     
-    # Generate the PNG
+    # Generate PNG of plot
     png(outfile)
     return(list(
       src = "documents/TCU_overview.png",