+++ server/server_plot.R
@@ -2,14 +2,14 @@
 
 observeEvent({
   
-  c(input$tabs, input$radio_mm, input$input_files)
+  c(input$tabs, input$radio_limit, input$input_files)
 },
 {
   if(is.null(LDF)) {
     return(NULL)
   }
   
-  p <- plotSubTrend(LDF, input$radio_mm[1])
+  p <- plotSubTrend(LDF, input$radio_limit[1])
   
   output$plot <- renderPlot({
     plot(p)