+++ ui/tab_plot.R
@@ -1,8 +1,8 @@
 # TCU trend analysis - tab plot
 
-tabPanel("Plot", value = "plot.tab",
+tabPanel("Plot", value = "plot",
          
-         conditionalPanel(condition = "output.file_uploaded & input.tabs == 'plot.tab'",
+         conditionalPanel(condition = "output.file_uploaded", # & input.tabs == 'plot'
                           plotOutput("plot")),
          
 )