+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2020-07-22
+
+### Added
+- Data validation steps
+
+### Changed
+- Updated README file
+- Changed radio button name into 'limits'
+- Edit text on main tab
+- Made irreleveant components invisble when switching tabs
+
 ## [0.0.2] - 2020-07-22
 
 ### Added